+++ src/Parallel-Solvers/Parallel-Explicit/Dynamic_Elastic_Solver/force_elastic.cpp
@@ -839,7 +839,7 @@ void FEA_Module_Dynamic_Elasticity::compute_stiffness_gradients(const_host_vec_a
         }
   }
 
-  for (int cycle = 0; cycle < last_time_step+1; cycle++) {
+  for (unsigned long cycle = 0; cycle < last_time_step+1; cycle++) {
     //compute timestep from time data
     global_dt = time_data[cycle+1] - time_data[cycle];
